pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
discord = "^1.7.3"
discord-py-slash-command = "^3.0.3"
google = "^3.0.0"
beautifulsoup4 = "^4.10.0"
Flask = "^2.0.3"
"quickchart.io" = "^1.0.1"
DateTime = "^4.4"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"